# api/openapi.yaml
openapi: 3.1.0
info:
  title: hermes-server
  description: Hermes server
  license:
    name: ''
  version: 0.1.0
paths:
  /api/v1/channel:
    post:
      tags:
      - channel
      operationId: create_channel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelPayload'
        required: true
      responses:
        '200':
          description: new channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          description: invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
  /api/v1/channel/{channel}:
    get:
      tags:
      - channel
      operationId: get_channel
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: accepted changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
    delete:
      tags:
      - channel
      operationId: delete_channel
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '201':
          description: accepted deletion
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
    patch:
      tags:
      - channel
      operationId: patch_channel
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchChannelPayload'
        required: true
      responses:
        '201':
          description: accepted changes
        '400':
          description: invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
  /api/v1/channel/{channel}/message:
    post:
      tags:
      - channel
      - message
      operationId: create_message
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessagePayload'
        required: true
      responses:
        '200':
          description: new message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '429':
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
  /api/v1/channel/{channel}/message/{message}:
    get:
      tags:
      - channel
      - message
      operationId: get_message
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: message
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '429':
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
    delete:
      tags:
      - channel
      - message
      operationId: delete_message
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: message
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '201':
          description: accepted deletion
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '429':
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
    patch:
      tags:
      - channel
      - message
      operationId: patch_message
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: message
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMessagePayload'
        required: true
      responses:
        '201':
          description: accepted update message
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '429':
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
  /api/v1/channel/{channel}/messages:
    get:
      tags:
      - channel
      - message
      operationId: list_messages
      parameters:
      - name: channel
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: accepted deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageQueryResult'
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '429':
          description: too many request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
...
components:
  schemas:
    ApplicationError:
      type: object
      required:
      - code
      - message
      - target
      - details
      properties:
        code:
          type: integer
          format: int32
          minimum: 0
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        innererror:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/InnerError'
        message:
          type: string
        target:
          type: string
    Channel:
      type: object
      required:
      - id
      - kind
      - name
      properties:
        category:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/ChannelKind'
        name:
          type: string
        server_id:
          type:
          - string
          - 'null'
          format: uuid
    ChannelKind:
      type: string
      enum:
      - text
      - voice
      - dm
    CreateChannelPayload:
      type: object
      required:
      - name
      - kind
      - server_id
      properties:
        category:
          type:
          - string
          - 'null'
        kind:
          $ref: '#/components/schemas/ChannelKind'
        name:
          type: string
        server_id:
          type: string
          format: uuid
...
    ErrorDetail:
      type: object
      required:
      - code
      - target
      - message
      properties:
        code:
          type: integer
          format: int32
          minimum: 0
        message:
          type: string
        target:
          type: string
    InnerError:
      type: object
      required:
      - trace
      properties:
        trace:
          type: array
          items:
            type: string
    Message:
      type: object
      required:
      - id
      - channel_id
      - content
      - created_at
      properties:
        channel_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
        edited_at:
          type:
          - string
          - 'null'
          format: date-time
        id:
          type: string
          format: uuid
        user_id:
          type:
          - string
          - 'null'
          format: uuid
    MessageQueryResult:
      type: object
      required:
      - results
      - count
      properties:
        count:
          type: integer
          minimum: 0
        cursor:
          type:
          - string
          - 'null'
        results:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    PatchChannelPayload:
      type: object
      properties:
        category:
          type:
          - string
          - 'null'
        name:
          type:
          - string
          - 'null'
    PatchMessagePayload:
      type: object
      required:
      - content
      properties:
        content:
          type: string
...
    PostMessagePayload:
      type: object
      required:
      - content
      properties:
        content:
          type: string
...